requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.0
hiredis==2.0.0
mongoengine==0.24.0
packaging==21.3
ptvsd==4.3.2
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.9
pytz==2021.3
redis==4.3.1
simplejson==3.17.6
sqlparse==0.2.4
toml==0.10.2
tzdata==2022.1
uuid==1.30
wrapt==1.14.1
gunicorn==20.1.0
